updated DS Player
