+++ package.json
@@ -39,7 +39,7 @@
     "@newrelic/newrelic-capacitor-plugin": "^1.3.1",
     "@nuxt/image": "^1.0.0-rc.1",
     "@nuxtjs/supabase": "^1.0.2",
-    "@nypublicradio/nypr-design-system-vue3": "^2.0.69",
+    "@nypublicradio/nypr-design-system-vue3": "^2.0.70",
     "@vuelidate/core": "^2.0.3",
     "@vuelidate/validators": "^2.0.3",
     "capacitor-native-settings": "^5.0.0",
